+++ cmd/varlogsn/cli.go
@@ -38,7 +38,7 @@ func newStartCommand() *cli.Command {
 			flagStorageNodeID.StringFlag(false, types.StorageNodeID(1).String()),
 			flagListen.StringFlag(false, "127.0.0.1:9091"),
 			flagAdvertise.StringFlag(false, ""),
-			flagBallastSize.StringFlag(false, "1G"),
+			flagBallastSize.StringFlag(false, storagenode.DefaultBallastSize),
 
 			// volumes
 			flagVolumes.StringSliceFlag(true, nil),
@@ -63,9 +63,10 @@ func newStartCommand() *cli.Command {
 			flagStorageL0StopWritesThreshold.IntFlag(false, storage.DefaultL0StopWritesThreshold),
 			flagStorageLBaseMaxBytes.StringFlag(false, units.ToByteSizeString(storage.DefaultLBaseMaxBytes)),
 			flagStorageMaxOpenFiles.IntFlag(false, storage.DefaultMaxOpenFiles),
-			flagStorageMemTableSize.StringFlag(false, units.ToByteSizeString(storage.DefaultMemTableSize)),
+			flagStorageMemTableSize,
 			flagStorageMemTableStopWritesThreshold.IntFlag(false, storage.DefaultMemTableStopWritesThreshold),
 			flagStorageMaxConcurrentCompaction.IntFlag(false, storage.DefaultMaxConcurrentCompactions),
+			flagStorageMetricsLogInterval,
 			flagStorageVerbose.BoolFlag(),
 
 			flagLogDir.StringFlag(false, ""),